soc/intel_adsp: Add hardware race workaround to cavstool
On cAVS 1.8 (specifically) there seems to be a propagation delay on
the IPC registers.  Hitting the TDA register to signal DONE too soon
after clearing the interrupt via TDR can cause the interrupt to be
dropped.  Merely polling for it to read back correctly isn't
sufficient, we need an actual sleep here.

(The behavior that a message won't send while an existing message is
in progress is actually a hardware feature that is new with 1.8.  My
guess is it's a little glitchy in its first version.)

(cherry picked from commit 005e12bdacf35c62123b37ebca287f1f9f2f6236)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 005e12bdacf35c62123b37ebca287f1f9f2f6236
Change-Id: Iab72734396d203435e55cf270eb7c95efe2a996e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3500455
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>