soc/intel_adsp: DMA Stability fix for cavstool
There is a hardware bug with stream reset, it won't stop the stream.
I thought it was limited to just the early versions, but it turns out
that 2.5 has it too in some variants.  Which is frustrating, because
the TGL chromebook I have doesn't like the START clear.  But all
systems work with an extra delay between them, so do that.

(cherry picked from commit 140837ce3ac6ba6ac01d804ebd40c8aaf00a3011)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 140837ce3ac6ba6ac01d804ebd40c8aaf00a3011
Change-Id: I17f8201bdb3cd85ed71b09552fc1e7fe39bf0ab3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3416814
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>