samples: grove_display: Exclude ip_k66f from sample build
The ip_k66f doesn't have I2C_0 status set to OKAY so the build of
this sample generates warnings in the nightly build. Root problem
is that the groove display is not ported to the device tree so
there isn't a standard way of filtering on the display being
part of the platform.

fixes #41523

(cherry picked from commit 96599830cf9eba2239a45008142831e93df9a6ec)

Signed-off-by: David Leach <david.leach@nxp.com>
GitOrigin-RevId: 96599830cf9eba2239a45008142831e93df9a6ec
Change-Id: Ifa68e14c29a18cb6ca70bea730c9af6126a8eaa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3427863
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>