drivers: pwm: add Renesas R-Car driver
Add R-Car Gen3 PWM driver.

Clock diviser is automatically adjusted according to requested period
and duty-cycle in order to obtain as much accuracy as possible.
Indeed, in order to improve PWM accurancy, the PWM clock has to fit
the requested period. So use the given period_cycle to define if the
clock as to be adapted. In such case, increase/decrease the clock
diviser to adapt the period_cycle and be sure that it fits into the
10 bits counter of the PWM controller.

Tested on H3ULCB on pwm0 and pwm4.

(cherry picked from commit 1d2752f4ee506efdb9b67a166a9037347be4fc37)

Original-Signed-off-by: Pierre Marzin <pierre.marzin@iot.bzh>
GitOrigin-RevId: 1d2752f4ee506efdb9b67a166a9037347be4fc37
Change-Id: Ibe87a47045ff388a699e2f649f5fb7c4cf35d979
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3882762
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>