pm: Clear the forced flag in the error path
Now that the forced flag is cleared with the kernel locked in the
suspend path, we need to clear it out when the suspend process fails.

This is being done before jump to the end label because in the
successful path the flag is already cleared.

(cherry picked from commit 350014d3993d8c59e8e2d5f20041deb707404129)

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: 350014d3993d8c59e8e2d5f20041deb707404129
Change-Id: I010efeaa7e0ce118401b805c45f21d28d7f51c26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3429334
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>