samples: sensor: th02: fix grove lcd usage
The sample enabled the Grove LCD by default in prj.conf, and in the
README it is clearly stated that the sample is intended to work using
the LCD module. This patch remove the unnecessary ifdeffery and obtains
a reference to the display at compile time.

(cherry picked from commit 9bbdc2486473b9a70528e40b562430d131150cd3)

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 9bbdc2486473b9a70528e40b562430d131150cd3
Change-Id: Ic5120407e392a5922ab5ecebb0f71732691ab1d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3484578
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>