Bluetooth: Audio: Remove bidirectional audio streams
Audio streams as defined by the BAP spec does not
support bidirectional audio streams. This commit
updates the API and implementation to match that.

The use a bidirectional CIS with 2 audio stream will
be added in a future commit.

This removes the _IN_ and _OUT_ and _INOUT_ QOS
values, as well as the direction of the codec QOS
struct.

To keep direction for internal use, the direction
has been added to the endpoint struct.

(cherry picked from commit b8806372318b37e5c4dfd5adb99141696677cb99)

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: b8806372318b37e5c4dfd5adb99141696677cb99
Change-Id: Ib753d38c5b7e4e71f53a41ba3db96b65a501339b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3519071
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>