test: settings: functional: fix missing test_main
The functional tests for settings subsystem regressed with the
introduction of the weak test_main function. This was caused by
Issue #19582 since the common code was being linked as a library.

Update the logic of the common code to directly add the sources
to the `app` target. This resolved both Issue #19582 and #42396

(cherry picked from commit 7f90587e7e4dc746b487c9021b612fb082f86037)

Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: 7f90587e7e4dc746b487c9021b612fb082f86037
Change-Id: I1db4d4b5afeafe779f15be877c157f4c3eafb617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3458926
Tested-by: CopyBot Service Account <copybot.service@gmail.com>