lib: os: crc: Rework the crc16() implementation
As described in #42403, there was an issue with the existing crc16_ansi()
implementation, since it was not calculating the CRC-16-ANSI (aka
CRC-16-MODBUS). This is because the  existing crc16() function only
supported non-reflected input and output (and the CRC-16-ANSI requires
reflection on both) and also it did not seem to support correctly inial
seeds different from 0x0000 (and, again, the CRC-16-ANSI requires 0xffff
as an initial seed).

This commit replaces the existing crc16() with a functional pair,
crc16() and crc16_reflect(), that also work with any poly, any initial seed
and allow to select whether reflection is performed.
It also adapts crc16_ansi() so that it actually returns the correct CRC.

Fixes #42403.

(cherry picked from commit 7f6524949ae8690049352951ec16b7e0a459fb36)

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
GitOrigin-RevId: 7f6524949ae8690049352951ec16b7e0a459fb36
Change-Id: I10925834f981d8e1aab9d70b1d03710cf63af96c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3442015
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>