Bluetooth: PACS: Refactor PAC location read/write
Refactor the PAC location read and write. Instead
of storing the location in the service, the
location is now stored in the application, and
is retrieved by the service via callbacks.

Similarly, if a client writes the location, this
request is being sent to the application.

(cherry picked from commit 18466530abd4a337fec18867cc08bc20acb910e9)

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 18466530abd4a337fec18867cc08bc20acb910e9
Change-Id: Ice586fa93a3ff225803db97c277bbf5fb1956c72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3520120
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>