arm64: switch to the IRQ stack during ISR execution
Avoid executing ISRs using the thread stack as it might not be sized
for that. Plus, we do have IRQ stacks already set up for us.

The non-nested IRQ context is still (and has to be) saved on the thread
stack as the thread could be preempted.

The irq_offload case is never nested and always invoked with the
sched_lock held so it can be simplified a bit.

(cherry picked from commit 34d425fbe53fdf09c25ed07d135ba137bd34ab7e)

Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
GitOrigin-RevId: 34d425fbe53fdf09c25ed07d135ba137bd34ab7e
Change-Id: I813751ff32488e1b774c5bdaa317a7d73f67df0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3480270
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>