lib/os: clarify that the winstream object does not track reads
Clarify that the reader MUST track the last sequence number read and
that the 'start' offset is not involved in that.

Some pre-conceptions + the sentence "the current state of the stream"
threw me off for a while. Thanks Andy for explaining this to me.

(cherry picked from commit 49900010132b9f905d59ea7eb3c9fae462e072bf)

Signed-off-by: Marc Herbert <marc.herbert@intel.com>
GitOrigin-RevId: 49900010132b9f905d59ea7eb3c9fae462e072bf
Change-Id: If89602a057bf43048fe0b9946fa76dd1ab4dac10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3480258
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>