driver: i2c: npcx: don't print the err log when SMBST is zero
In the I2C ISR, it prints the error message when SMBST is set to an
unexpected value. However, we found a spurious interrupt which caused
the SoC to enter the ISR with SMBST=0. Because the spurious interrupt
will not break the I2C operation, this commit limits the error log to
print if SMBST is not equivalent to 0 to prevent a false alert.

(cherry picked from commit 0bbf394ced742a448d24f9763709b0618651dadb)

Signed-off-by: Andrew McRae <amcrae@google.com>
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
GitOrigin-RevId: 0bbf394ced742a448d24f9763709b0618651dadb
Change-Id: I015e541edbb41b66a7d4b3d235dcdf6cc17a7d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3484558
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>