Bluetooth: Audio: ASCS: do not fail on preferred value check
bt_audio_valid_stream_qos() is used to validate the QoS parameters
set by a client. It did a check of a preferred setting which is
ok to exceed (mandatory in some cases).
This change removes the statement that causes the check to fail,
but keeps the check for debug purposes.

Fixes #43359

Signed-off-by: Casper Bonde <casper_bonde@bose.com>
GitOrigin-RevId: 0874ff5cfd90a881b63a611109ae00691fa97dcd
Change-Id: Idf7594660b1c6af0fd6aabd1809c9a4694ff092e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3552861
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>