Bluetooth: Host: Fix L2CAP collision mitigation
There were some errors in the detection of ECRED connection collisions,
so the retry was triggered incorrectly.

The number of channels requested in the retry was wrong in some cases,
using the number of channels from the received request instead of the
sent request.

Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>
GitOrigin-RevId: 1069d5fe56f0b6ea3f08de4ba93b5c5402c0c01d
Change-Id: I43acb1da2da60c5cf733e2f3abc9fd33b70fcc87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3641954
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>