ci: bluetooth-tests: Trigger on workflow changes
The 'bluetooth-tests' workflow is currently written such that it
only triggers on the pull requests that make changes to the files under
specific paths, which do not include the workflow itself.

This allows the workflow changes to be merged without being tested in
the CI and may lead to CI breakages when such changes contain errors.

This commit adds the workflow itself to the list of trigger paths so
that any changes made to the 'bluetooth-tests.yaml' and
'bluetooth-tests-publish.yaml' workflows are tested in the CI before
merging.

(cherry picked from commit ff5591e68939eb12e2592d25203764ee90e3836d)

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: ff5591e68939eb12e2592d25203764ee90e3836d
Change-Id: I492a477db49360f728a735623db921399e899ade
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3504559
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>